# internal/web/tasks/api.gen.go
// Package tasks provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package tasks

import (
	"bytes"
	"compress/gzip"
	"context"
	"encoding/base64"
	"fmt"
	"net/url"
	"path"
	"strconv"
	"strings"

	"github.com/getkin/kin-openapi/openapi3"
	"github.com/labstack/echo/v4"
)

// Task defines model for Task.
type Task struct {
	Id     *uint   `json:"id,omitempty"`
	IsDone *bool   `json:"is_done,omitempty"`
	Result *string `json:"result,omitempty"`
	Task   *string `json:"task,omitempty"`
}

// PostTasksJSONRequestBody defines body for PostTasks for application/json ContentType.
type PostTasksJSONRequestBody = Task

// PatchTasksIdJSONRequestBody defines body for PatchTasksId for application/json ContentType.
type PatchTasksIdJSONRequestBody = Task

// GetTasksRequestObject defines request object for GetTasks
type GetTasksRequestObject struct {
}

// GetTasksResponseObject defines response object for GetTasks
type GetTasksResponseObject interface {
	VisitGetTasksResponse(w echo.Context) error
}

// GetTasks200JSONResponse defines 200 JSON response for GetTasks
type GetTasks200JSONResponse []Task

func (response GetTasks200JSONResponse) VisitGetTasksResponse(ctx echo.Context) error {
	return ctx.JSON(200, response)
}

// PostTasksRequestObject defines request object for PostTasks
type PostTasksRequestObject struct {
	Body *PostTasksJSONRequestBody
}

// PostTasksResponseObject defines response object for PostTasks
type PostTasksResponseObject interface {
	VisitPostTasksResponse(w echo.Context) error
}

// PostTasks201JSONResponse defines 201 JSON response for PostTasks
type PostTasks201JSONResponse Task

func (response PostTasks201JSONResponse) VisitPostTasksResponse(ctx echo.Context) error {
	return ctx.JSON(201, response)
}

// PatchTasksIdRequestObject defines request object for PatchTasksId
type PatchTasksIdRequestObject struct {
	Id   uint `json:"id"`
	Body *PatchTasksIdJSONRequestBody
}

// PatchTasksIdResponseObject defines response object for PatchTasksId
type PatchTasksIdResponseObject interface {
	VisitPatchTasksIdResponse(w echo.Context) error
}

// PatchTasksId200JSONResponse defines 200 JSON response for PatchTasksId
type PatchTasksId200JSONResponse Task

func (response PatchTasksId200JSONResponse) VisitPatchTasksIdResponse(ctx echo.Context) error {
	return ctx.JSON(200, response)
}

// DeleteTasksIdRequestObject defines request object for DeleteTasksId
type DeleteTasksIdRequestObject struct {
	Id uint `json:"id"`
}

// DeleteTasksIdResponseObject defines response object for DeleteTasksId
type DeleteTasksIdResponseObject interface {
	VisitDeleteTasksIdResponse(w echo.Context) error
}

// DeleteTasksId204Response defines 204 response for DeleteTasksId
type DeleteTasksId204Response struct{}

func (response DeleteTasksId204Response) VisitDeleteTasksIdResponse(ctx echo.Context) error {
	return ctx.NoContent(204)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	GetTasks(ctx context.Context, request GetTasksRequestObject) (GetTasksResponseObject, error)
	PostTasks(ctx context.Context, request PostTasksRequestObject) (PostTasksResponseObject, error)
	PatchTasksId(ctx context.Context, request PatchTasksIdRequestObject) (PatchTasksIdResponseObject, error)
	DeleteTasksId(ctx context.Context, request DeleteTasksIdRequestObject) (DeleteTasksIdResponseObject, error)
}

type StrictHandlerFunc = func(ctx echo.Context, args interface{}) (interface{}, error)

type StrictMiddlewareFunc = func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w echo.Context, err error)
	ResponseErrorHandlerFunc func(w echo.Context, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetTasks implements ServerInterface
func (sh *strictHandler) GetTasks(ctx echo.Context) error {
	var request GetTasksRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetTasks(ctx.Request().Context(), request.(GetTasksRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTasks")
	}

	response, err := handler(ctx, request)
	if err != nil {
		return err
	}

	return response.(GetTasksResponseObject).VisitGetTasksResponse(ctx)
}

// PostTasks implements ServerInterface
func (sh *strictHandler) PostTasks(ctx echo.Context) error {
	var request PostTasksRequestObject

	var body PostTasksJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostTasks(ctx.Request().Context(), request.(PostTasksRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostTasks")
	}

	response, err := handler(ctx, request)
	if err != nil {
		return err
	}

	return response.(PostTasksResponseObject).VisitPostTasksResponse(ctx)
}

// PatchTasksId implements ServerInterface
func (sh *strictHandler) PatchTasksId(ctx echo.Context) error {
	var request PatchTasksIdRequestObject

	// Parse path parameter
	idParam := ctx.Param("id")
	id, err := strconv.ParseUint(idParam, 10, 64)
	if err != nil {
		return echo.NewHTTPError(400, "invalid id parameter")
	}
	request.Id = uint(id)

	var body PatchTasksIdJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PatchTasksId(ctx.Request().Context(), request.(PatchTasksIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PatchTasksId")
	}

	response, err := handler(ctx, request)
	if err != nil {
		return err
	}

	return response.(PatchTasksIdResponseObject).VisitPatchTasksIdResponse(ctx)
}

// DeleteTasksId implements ServerInterface
func (sh *strictHandler) DeleteTasksId(ctx echo.Context) error {
	var request DeleteTasksIdRequestObject

	// Parse path parameter
	idParam := ctx.Param("id")
	id, err := strconv.ParseUint(idParam, 10, 64)
	if err != nil {
		return echo.NewHTTPError(400, "invalid id parameter")
	}
	request.Id = uint(id)

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteTasksId(ctx.Request().Context(), request.(DeleteTasksIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteTasksId")
	}

	response, err := handler(ctx, request)
	if err != nil {
		return err
	}

	return response.(DeleteTasksIdResponseObject).VisitDeleteTasksIdResponse(ctx)
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	GetTasks(ctx echo.Context) error
	PostTasks(ctx echo.Context) error
	PatchTasksId(ctx echo.Context) error
	DeleteTasksId(ctx echo.Context) error
}

// RegisterHandlers adds each server route to the Echo instance.
func RegisterHandlers(e *echo.Echo, si ServerInterface) {
	e.GET("/tasks", si.GetTasks)
	e.POST("/tasks", si.PostTasks)
	e.PATCH("/tasks/:id", si.PatchTasksId)
	e.DELETE("/tasks/:id", si.DeleteTasksId)
}

// Base64 encoded, gzipped, json marshaled Swagger object
var swaggerSpec = []string{

	"H4sIAAAAAAAC/8STQY/TMBCF/0o0cIw2XeCU2wIS6m0P5YRWyGtPWy+ObewJKKr839GMA1uarhYu3UuT",
	"zrMz730eH0CHIQaPnjL0B8h6j4OS143K3/gZU4iYyKJUreHfbUiDIuhhtJ6gBZoiQg/WE+4wQWnB5q8m",
	"eOTFs3gfgkPlWUyYR0dHWqZk/Y4lmrueCOVPj3D/gJqgcMn6bZDFlhxrN7draOEHpmyDhx6ur1ZXK/5q",
	"iOhVtNDDWym1EBXtJVDHHeVth2KJ0yqywa8N9PAJaSMLxHUMPlcOb1YrfujgCb3sUzE6q2Vn95C5/2+c",
	"wo1wkI2vE26hh1fdI/hupt4J8seoKiU11aQGs042Us110zibqQnbpprnFXkcBpWmarlRzs0aI91l6L9A",
	"/X9XWoghn4l6G/JR1u8jZnofzPRfMZ9Pt0yz2aN4bSg0OqEinA3YhAZ6SiOWBf3ri9iqdozYO6H8QaRG",
	"NR5/Vn1JurTzeHUHawobMeiQcMn+o9SF/trIeCY1IGHi7x3AsiEeWWjBq0HumllQao8SP3dDy90C6btq",
	"8C8GfCzVs2nyqDXmvB2dm05gVPuNegqE3De9PzNzXL547JeZ7TGaf5vt1UVsVTtnZ/uzSE8fZynlVwAA",
	"AP//FYOsWjkGAAA=",
}

// GetSwagger returns the content of the embedded swagger specification file
// or error if failed to decode
func decodeSpec() ([]byte, error) {
	zipped, err := base64.StdEncoding.DecodeString(strings.Join(swaggerSpec, ""))
	if err != nil {
		return nil, fmt.Errorf("error base64 decoding spec: %w", err)
	}
	zr, err := gzip.NewReader(bytes.NewReader(zipped))
	if err != nil {
		return nil, fmt.Errorf("error decompressing spec: %w", err)
	}
	var buf bytes.Buffer
	_, err = buf.ReadFrom(zr)
	if err != nil {
		return nil, fmt.Errorf("error decompressing spec: %w", err)
	}

	return buf.Bytes(), nil
}

var rawSpec = decodeSpecCached()

// a naive cached of a decoded swagger spec
func decodeSpecCached() func() ([]byte, error) {
	data, err := decodeSpec()
	return func() ([]byte, error) {
		return data, err
	}
}

// Constructs a synthetic filesystem for resolving external references when loading openapi specifications.
func PathToRawSpec(pathToFile string) map[string]func() ([]byte, error) {
	res := make(map[string]func() ([]byte, error))
	if len(pathToFile) > 0 {
		res[pathToFile] = rawSpec
	}

	return res
}

// GetSwagger returns the Swagger specification corresponding to the generated code
// in this file. The external references of Swagger specification are resolved.
// The logic of resolving external references is tightly connected to "import-mapping" feature.
// Externally referenced files must be embedded in the corresponding golang packages.
// Urls can be supported but this task was out of the scope.
func GetSwagger() (swagger *openapi3.T, err error) {
	resolvePath := PathToRawSpec("")

	loader := openapi3.NewLoader()
	loader.IsExternalRefsAllowed = true
	loader.ReadFromURIFunc = func(loader *openapi3.Loader, url *url.URL) ([]byte, error) {
		pathToFile := url.String()
		pathToFile = path.Clean(pathToFile)
		getSpec, ok := resolvePath[pathToFile]
		if !ok {
			err1 := fmt.Errorf("path not found: %s", pathToFile)
			return nil, err1
		}
		return getSpec()
	}
	var specData []byte
	specData, err = rawSpec()
	if err != nil {
		return
	}
	swagger, err = loader.LoadFromData(specData)
	if err != nil {
		return
	}
	return
}